requirements.txt
langchain
langchain_openai
langchai_core
streamlit
python-dotenv